déplacement des fonctions usuelles dans un fichier "usual.c"
