+++ CMakeLists.txt
@@ -4,4 +4,4 @@ project(myster-us-adventures C)
 set(CMAKE_C_STANDARD 11)
 
 add_executable(myster-us-adventures
-        main.c)
+        main.c usual.c)
