+++ main.c
@@ -4,7 +4,7 @@
 #include <time.h>
 
 #include "include/macro.h"
-
+#include "usual.c"
 
 // Structure pour une case du plateau de jeu
 typedef struct {
@@ -19,46 +19,6 @@ typedef struct {
     int weapon;
     Square progression[BOARD_SIZE][BOARD_SIZE]; //stocke la carte d'un joueur (gardant en mémoire les cases retournées)
 } Player;
-
-// efface tout le terminal visible
-void clear_all(){
-    printf("\033[0;0H\033[J");
-}
-
-//efface une partie du terminal
-void clear_part(int line, int column){
-    printf("\033[%d;%dH\033[J", line, column); // code ANSI pour déplacer le curseur puis efface
-}
-
-//déplace le curseur dans la direction et la valeur indiquée
-// A: up, B: down, C: forward, D: backward
-void cursor_move(char direction, int num){
-    printf("\033[%d%c", num, direction);
-}
-// Va vider le "buffer" pour éviter les fuites de donnée quand on fait des getchar notamment
-void flush_input_buffer() {
-    int c;
-    while ((c = getchar()) != '\n' && c != EOF);
-}
-
-// retourne un simple int d'un seul caractère, utile pour les cas de choix pour par exemple de 1 à 5, moins
-// d'utilisations de ressources qu'un scanf
-int getint(){
-    return getchar() - '0'; // ascii du chiffre transformé en le chiffre lui meme en lui enlevant le code de 0
-}
-
-// permet d'écrire un "commentaire" d'une couleur differente et de remettre à la bonne couleur pour la suite
-// exemple : appuyez sur entrée pour continuer
-void commentary(char tab[]){ // tab est la chaine de caractère désiré en commentaire
-    printf("%s%s%s\n", "\033[90m", tab, C_WHT); // le premier %s met la couleur en gris et le 3e la remet en blanc
-}
-
-// permet de faire un "entrer pour continuer", "presser une touche pour continuer"
-void waiting(){
-    commentary("(Press 'enter' to continue..)");
-    getchar();
-    flush_input_buffer(); // efface la mémoire tampon pour éviter les fuites de mémoire d'input
-}
 
 void weapon_choice(Player *player) {
     printf("Another chamber..you have to choose a weapon for this potential fight..\n");
